+++ Ems/src/main/java/com/example/Ems/department/DepartmentService.java
@@ -2,8 +2,12 @@ package com.example.Ems.department;
 
 import com.example.Ems.Employee.Employee;
 import com.example.Ems.Employee.EmployeeResponse;
+import com.example.Ems.configuration.NotFoundInDatabaseException;
+import com.example.Ems.project.Project;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -29,8 +33,10 @@ public class DepartmentService {
         return departmentResponses;
     }
 
-    public Department getDepartmentById(Integer id) {
-        return departmentRepo.findById(id).get();
+    public ResponseEntity<?>findById(Integer id) throws NotFoundInDatabaseException {
+        Department department = departmentRepo.findById(id).orElseThrow(()-> new NotFoundInDatabaseException("department not found"));
+        DepartmentResponse response = mapToResponse(department);
+        return ResponseEntity.status(HttpStatus.OK).body(response);
     }
 
     public DepartmentResponse addDepartment(DepartmentRequest departmentRequest) {
@@ -50,12 +56,25 @@ public class DepartmentService {
 
     }
 
-    public void deleteById(Integer id) {
+    public ResponseEntity<?> deleteById(Integer id)throws NotFoundInDatabaseException {
+
+        Department department =departmentRepo.findById(id).orElseThrow(()-> new NotFoundInDatabaseException("department not found"));
+        if(department == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("departmen not found");
+        }
         departmentRepo.deleteById(id);
+        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
 
-    public Department update(Department department) {
-        return departmentRepo.save(department);
+    public ResponseEntity <?>update( Integer id ,DepartmentRequest request )throws NotFoundInDatabaseException {
+        Department department = departmentRepo.findById(id).orElseThrow(()-> new NotFoundInDatabaseException("Project not found"));
+        department.setName(request.getName());
+        department.setDescription(request.getDescription());
+        department =  departmentRepo.save(department);
+        return ResponseEntity.status(HttpStatus.OK).body(department);
+    }
+    public DepartmentResponse mapToResponse(Department department) {
+        return  new DepartmentResponse();
     }
